+++ src/app/app-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { RjuneGameComponent } from './rjune-game/rjune-game.component';
+import { MainComponent } from './main/main.component';
 
 const route = (path: string, component: any) => ({ path, component });
-const routes: Routes = [route('rjune', RjuneGameComponent)];
+const routes: Routes = [route('factoryClicker', MainComponent)];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
