+++ src/app/app.component.html
@@ -8,13 +8,5 @@
 </style>
 
 <div class = "content">
-  <!-- <div>
-      <h1>{{this.title}}</h1>
-  </div>
-
-  <input [(ngModel)] = "currentLog" /> 
-  <button (click)="this.addLog()">Add log entry</button>
-  <app-logger [logs]="this.logs"></app-logger>
-  <button (click)="this.clearLogs()">Clear</button> -->
   <router-outlet></router-outlet>
 </div>
