+++ src/app/main/main.component.html
@@ -8,7 +8,7 @@
     }
 </style>
 
-<div class = "rjune-root">
+<div class = "factory-root">
     
     <p>Clicks: {{this.clickCount}}</p>
     <p>Level: {{this.currentLevel}}</p>
@@ -69,7 +69,7 @@
        <app-assembler [recipe]="this.currentOption" [resourceTransferer] = "this.resourceTransferer"></app-assembler>
    </div>
 
-   <div>
-    <h2> Science </h2>       
-</div>
+    <div>
+        <h2> Science </h2>       
+    </div>
 </div>
