+++ src/factoryClicker/RecipeResult.ts
