+++ src/factoryClicker/RequiredResourceInfo.ts
