+++ src/factoryClicker/ResourceType.ts
